refactor: rename userModule -> adminModule
