+++ src/api/database/db.module.ts
@@ -1,10 +1,12 @@
 // TODO: Think of a way to merge prisma and supabase services into a singular database service
-import { Module } from "@nestjs/common";
+import { Global, Module } from "@nestjs/common";
+import { ConfigService } from "@nestjs/config";
+import { PrismaService } from "./prisma.service";
+import { SupabaseService } from "./supabase.service";
 
+@Global()
 @Module({
-    imports: [],
-    controllers: [],
-    providers: [],
-    exports: [],
+    providers: [PrismaService, SupabaseService, ConfigService],
+    exports: [PrismaService, SupabaseService],
 })
 export class DatabaseModule {}
