+++ src/api/project/project.module.ts
@@ -4,7 +4,7 @@ import { APP_GUARD } from "@nestjs/core";
 import { JwtService } from "@nestjs/jwt";
 import { PasswordService } from "src/api/auth/pwd.service";
 import { AuthGuard } from "src/shared/guards/auth.guard";
-import { SupabaseService } from "../../supabase/supabase.service";
+import { SupabaseService } from "../database/supabase.service";
 import { ProjectController } from "./project.controller";
 import { ProjectService } from "./project.service";
 import { PublicProjectController } from "./public.controller";
