+++ src/app.module.ts
@@ -1,18 +1,18 @@
 import { MiddlewareConsumer, Module, NestModule } from "@nestjs/common";
-import { UserModule } from "./api/admin/admin.module";
+import { AdminModule } from "./api/admin/admin.module";
 import { AuthModule } from "./api/auth/auth.module";
+import { DatabaseModule } from "./api/database/db.module";
 import { HealthModule } from "./api/health/health.module";
 import { ProjectModule } from "./api/project/project.module";
-import { PrismaModule } from "./prisma/prisma.module";
 import { HttpLoggerMiddleware } from "./shared/middlewares/logger";
 
 @Module({
     imports: [
         AuthModule,
-        UserModule,
-        PrismaModule,
+        AdminModule,
         ProjectModule,
         HealthModule,
+        DatabaseModule,
     ],
 })
 export class AppModule implements NestModule {
