+++ src/supabase/supabase.module.ts
@@ -1,6 +1,6 @@
 import { Global, Module } from "@nestjs/common";
 import { ConfigService } from "@nestjs/config";
-import { PrismaService } from "../prisma/prisma.service";
+import { PrismaService } from "../api/database/prisma.service";
 
 @Global()
 @Module({
